requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.3.0
exceptiongroup==1.3.0
fastapi==0.103.2
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
pydantic==2.5.3
pydantic-core==2.14.6
pymongo==4.7.3
python-dotenv==0.21.1
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
zipp==3.15.0
